api/requirements.txt
flask-cors==3.0.10
gphoto2==2.3.0
pillow==8.4.0